Issue 83: Proposal for pravega-samples repository reorganization.
