+++ scenarios/anomaly-detection/src/main/java/io/pravega/anomalydetection/ApplicationMain.java
