+++ scenarios/anomaly-detection/src/main/java/io/pravega/anomalydetection/event/pipeline/EventStateMachineMapper.java
