+++ scenarios/anomaly-detection/src/main/java/io/pravega/anomalydetection/event/state/EventStateMachine.java
