+++ scenarios/anomaly-detection/src/main/resources/Network-Anomaly.png
