--- standalone-examples/bin/.gitignore
@@ -1,2 +1,0 @@
-/io/
-/logback.xml
